vbox_create.sh: Do not use cache-dir unless it already exists.

+++ vbox_create.sh
@@ -41,18 +41,15 @@ function download_VM_files {
 
   ROM=gpxe-1.0.1-80861004.rom
   CACHEDIR=~/bcpc-cache
-  if [[ ! -d $CACHEDIR ]]; then
-      mkdir $CACHEDIR
-  fi
 
-  if [[ ! -f  $ROM ]]; then
+  if [[ ! -f $ROM ]]; then
       if [[ -f $CACHEDIR/$ROM ]]; then
 	  cp $CACHEDIR/$ROM .
       else
 	  $CURL -o gpxe-1.0.1-80861004.rom "http://rom-o-matic.net/gpxe/gpxe-1.0.1/contrib/rom-o-matic/build.php" -H "Origin: http://rom-o-matic.net" -H "Host: rom-o-matic.net" -H "Content-Type: application/x-www-form-urlencoded" -H "Accept: text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8" -H "Referer: http://rom-o-matic.net/gpxe/gpxe-1.0.1/contrib/rom-o-matic/build.php" -H "Accept-Charset: ISO-8859-1,utf-8;q=0.7,*;q=0.3" --data "version=1.0.1&use_flags=1&ofmt=ROM+binary+%28flashable%29+image+%28.rom%29&nic=all-drivers&pci_vendor_code=8086&pci_device_code=1004&PRODUCT_NAME=&PRODUCT_SHORT_NAME=gPXE&CONSOLE_PCBIOS=on&BANNER_TIMEOUT=20&NET_PROTO_IPV4=on&COMCONSOLE=0x3F8&COMSPEED=115200&COMDATA=8&COMPARITY=0&COMSTOP=1&DOWNLOAD_PROTO_TFTP=on&DNS_RESOLVER=on&NMB_RESOLVER=off&IMAGE_ELF=on&IMAGE_NBI=on&IMAGE_MULTIBOOT=on&IMAGE_PXE=on&IMAGE_SCRIPT=on&IMAGE_BZIMAGE=on&IMAGE_COMBOOT=on&AUTOBOOT_CMD=on&NVO_CMD=on&CONFIG_CMD=on&IFMGMT_CMD=on&IWMGMT_CMD=on&ROUTE_CMD=on&IMAGE_CMD=on&DHCP_CMD=on&SANBOOT_CMD=on&LOGIN_CMD=on&embedded_script=&A=Get+Image"
 	      
       fi
-      if [[ ! -f $CACHEDIR/$ROM ]]; then
+      if [[ -d $CACHEDIR && ! -f $CACHEDIR/$ROM ]]; then
 	  cp $ROM $CACHEDIR/$ROM
       fi
   fi
@@ -67,7 +64,7 @@ function download_VM_files {
      #$CURL -o ubuntu-12.04-mini.iso http://archive.ubuntu.com/ubuntu/dists/precise/main/installer-amd64/current/images/netboot/mini.iso
 	  $CURL -o $ISO http://archive.ubuntu.com/ubuntu/dists/precise-updates/main/installer-amd64/current/images/raring-netboot/mini.iso
       fi
-      if [[ ! -f $CACHEDIR/$ISO ]]; then
+      if [[ -d $CACHEDIR && ! -f $CACHEDIR/$ISO ]]; then
 	  cp $ISO $CACHEDIR
       fi
   fi
@@ -83,7 +80,7 @@ function download_VM_files {
 	else
 	    $CURL -o precise-server-cloudimg-amd64-vagrant-disk1.box http://cloud-images.ubuntu.com/vagrant/precise/current/precise-server-cloudimg-amd64-vagrant-disk1.box
 	fi
-	if [[ ! -f $CACHEDIR/$BOX ]]; then
+	if [[ -d $CACHEDIR && ! -f $CACHEDIR/$BOX ]]; then
 	    cp $BOX $CACHEDIR
 	fi
     fi
